Create about component
